One tape to a file (#51)
`tapes.rs` crossed the 500-line wall and `speed_tape` crossed the
80-line one, so `scripts/check-structure.sh` — a hard CI gate — was
red.

The seam is the tape, not the line count. The speed tape and the
altitude tape share no geometry beyond the strip they run down: each
has its own scale, its own bands or bugs, its own boxes. What they do
share is the vertical extent both span and the pointed readout each
hangs at the pointer line, and that is what stays in `tapes.rs`. The
VSI travels with the altitude tape it reads against, which is where a
reader looking for it would go.

The groundspeed box comes out of `speed_tape` as `gs_box`, mirroring
the `tas_box` already beside it: the function drew both boxes and only
one of them had a name.

Both tapes end at the same y and hang their readout boxes from it, so
that value gets a name instead of appearing as a bare 335.0 in a
clamp, a min, and two box anchors.
