+++ sets/indicate-instrument-panels/src/pfd/tapes.rs
@@ -3,214 +3,31 @@
 //! Tape scaling from the G5 proportions: speed ±25 kt over the panel
 //! height (7.2 px/kt), altitude ±150 ft (1.2 px/ft), VSI ±1500 fpm full
 //! scale.
+//!
+//! Each tape draws from its own module, because the two share no
+//! geometry beyond the strip they run down. What stays here is exactly
+//! that shared part: the vertical extent both span, and the pointed
+//! readout each hangs at the pointer line.
 
 use indicate_instrument_scene::{
-    Anchor, PaintMode, Rgba8, SceneError, SceneWriter, nominal_text_ink_width, nominal_text_width,
+    Anchor, PaintMode, SceneError, SceneWriter, nominal_text_ink_width, nominal_text_width,
 };
-use indicate_instrument_state::AltitudeClass;
-use indicate_instrument_state::{GroupId, PanelData, Sig};
+use indicate_instrument_state::Sig;
+use indicate_instrument_symbology::{palette, safety, status_paint};
 
-use indicate_instrument_symbology::{fmt_label, palette, safety, status_paint};
+mod airspeed;
+mod altitude;
 
-use super::VSpeeds;
+pub use airspeed::speed_tape;
+pub use altitude::{altitude_tape, vsi};
 
-const PX_PER_KT: f32 = 7.2;
-/// Top of the airspeed tape. The true-airspeed box owns the strip above
-/// it: the box is opaque, so a tape that started at the frame edge
-/// would have its topmost gradation painted over rather than covered by
-/// a box beside it.
-const SPEED_TAPE_TOP: f32 = 25.0;
-const PX_PER_FT: f32 = 1.2;
+/// The pointer line: the height at which either tape reads its value.
 const CENTER_Y: f32 = 180.0;
 
-/// Left-edge airspeed tape with bands, readout, and the TAS and
-/// groundspeed boxes at its head and foot.
-pub fn speed_tape(
-    scene: &mut SceneWriter<'_>,
-    data: &PanelData,
-    v: Option<&VSpeeds>,
-    declutter: bool,
-) -> Result<(), SceneError> {
-    let ias = data.ias_kt;
-    scene.fill_color(palette::TAPE_BG)?;
-    scene.rect(
-        PaintMode::Fill,
-        0.0,
-        SPEED_TAPE_TOP,
-        90.0,
-        335.0 - SPEED_TAPE_TOP,
-    )?;
-
-    if ias.status.shows_value() {
-        if let Some(v) = v {
-            speed_bands(scene, ias.value, v)?;
-        }
-        scene.save()?;
-        scene.clip_rect(0.0, SPEED_TAPE_TOP, 90.0, 335.0 - SPEED_TAPE_TOP)?;
-        scene.stroke(palette::WHITE, 2.0)?;
-        scene.fill_color(palette::WHITE)?;
-        let lo = (((ias.value - 26.0) / 5.0) as i32).max(0);
-        let hi = ((ias.value + 26.0) / 5.0) as i32;
-        for step in lo..=hi {
-            let kt = step * 5;
-            let y = CENTER_Y - (kt as f32 - ias.value) * PX_PER_KT;
-            scene.line(78.0, y, 90.0, y)?;
-            if step % 2 == 0 {
-                let label = fmt_label!(8, "{kt}");
-                scene.text_attributed(
-                    GroupId::Air.to_u8(),
-                    70.0,
-                    y,
-                    20.0,
-                    Anchor::CENTER,
-                    label.as_str(),
-                )?;
-            }
-        }
-        scene.restore()?;
-    } else {
-        scene.fill_color(palette::GREY)?;
-        scene.text(45.0, 130.0, 16.0, Anchor::CENTER, "IAS")?;
-    }
-
-    // Pointed readout box, always drawn so `Missing` shows dashes.
-    let text = fmt_label!(8, "{:03}", libm::roundf(ias.value) as i32);
-    pointed_readout(
-        scene,
-        GroupId::Air.to_u8(),
-        ias,
-        text.as_str(),
-        &IAS_READOUT,
-    )?;
-
-    if !declutter {
-        trend_bar(scene, data)?;
-    }
-
-    tas_box(scene, data)?;
-
-    // Groundspeed box under the tape.
-    let gs = data.gs_kt;
-    let gs_text = fmt_label!(12, "GS {:.0}kt", gs.value);
-    status_paint::readout_box(
-        scene,
-        GroupId::Kinematics.to_u8(),
-        0.0,
-        335.0,
-        90.0,
-        25.0,
-        gs_text.as_str(),
-        palette::MAGENTA,
-        16.0,
-        gs.status,
-    )?;
-    Ok(())
-}
-
-/// How far ahead the trend cue reads, in seconds. The bar marks where
-/// the airspeed will be if the current rate holds, so the look-ahead is
-/// the whole meaning of its length and belongs beside it.
-const TREND_LOOK_AHEAD_S: f32 = 6.0;
-
-/// The airspeed trend bar, just outside the tape's inner edge: from the
-/// pointer line to where the airspeed will be after
-/// [`TREND_LOOK_AHEAD_S`] at the current rate.
-///
-/// Drawn only when the tape itself is showing a value — a trend beside
-/// dashes marks a change in a number the pilot cannot read. An absent
-/// rate draws nothing at all, because a zero-length bar would claim the
-/// airspeed is steady, which is a different statement from not knowing.
-fn trend_bar(scene: &mut SceneWriter<'_>, data: &PanelData) -> Result<(), SceneError> {
-    let trend = data.ias_trend_kt_s;
-    if !trend.status.shows_value() || !data.ias_kt.status.shows_value() {
-        return Ok(());
-    }
-    let reach = trend.value * TREND_LOOK_AHEAD_S * PX_PER_KT;
-    // The tip stops at the tape's own ends — which start below the
-    // true-airspeed box, not at the frame edge. Past them the bar would
-    // point at a speed the tape is not showing.
-    let tip = (CENTER_Y - reach).clamp(SPEED_TAPE_TOP, 335.0);
-    let (top, height) = if tip < CENTER_Y {
-        (tip, CENTER_Y - tip)
-    } else {
-        (CENTER_Y, tip - CENTER_Y)
-    };
-    if height <= 0.0 {
-        return Ok(());
-    }
-    scene.fill_color(palette::MAGENTA)?;
-    scene.rect(PaintMode::Fill, 90.0, top, 4.0, height)?;
-    Ok(())
-}
-
-/// True-airspeed box at the head of the tape, mirroring the groundspeed
-/// box at its foot. TAS is air data, so the box wears primary white
-/// where the kinematic-derived GS box wears magenta; an absent TAS (a
-/// source may supply IAS alone) shows this box's dashes and leaves the
-/// tape itself untouched.
-fn tas_box(scene: &mut SceneWriter<'_>, data: &PanelData) -> Result<(), SceneError> {
-    let tas = data.tas_kt;
-    let tas_text = fmt_label!(12, "TAS {:.0}kt", tas.value);
-    status_paint::readout_box(
-        scene,
-        GroupId::Air.to_u8(),
-        0.0,
-        0.0,
-        90.0,
-        SPEED_TAPE_TOP,
-        tas_text.as_str(),
-        palette::WHITE,
-        fitted_label_size(90.0, tas_text.as_str().chars().count(), 16.0),
-        tas.status,
-    )
-}
-
-/// Largest size, capped at `preferred`, whose nominal ink fits `width`.
-///
-/// A centered label in a box of fixed width overflows on both sides once
-/// its ink outstrips the box, and the frame edge is one of those sides:
-/// `TAS 113kt` at 16 units carries 121 units of ink into a 90-unit box,
-/// so the leading glyph paints off the panel entirely. The size follows
-/// the value's width instead, which is what the pointed readouts already
-/// do (DISP-02).
-fn fitted_label_size(width: f32, chars: usize, preferred: f32) -> f32 {
-    let ink = nominal_text_ink_width(preferred, chars);
-    if ink <= width {
-        return preferred;
-    }
-    preferred * width / ink
-}
-
-fn speed_bands(scene: &mut SceneWriter<'_>, ias: f32, v: &VSpeeds) -> Result<(), SceneError> {
-    let segs: [(f32, f32, Rgba8); 3] = [
-        (v.vs_kt, v.vno_kt, palette::BAND_GREEN),
-        (v.vno_kt, v.vne_kt, safety::BAND_CAUTION),
-        (v.vne_kt, v.vne_kt + 1000.0, safety::FAILURE_RED),
-    ];
-    for (lo, hi, color) in segs {
-        band_rect(scene, ias, lo, hi, 86.0, 4.0, color)?;
-    }
-    band_rect(scene, ias, v.vs0_kt, v.vfe_kt, 82.0, 4.0, palette::WHITE)?;
-    Ok(())
-}
-
-fn band_rect(
-    scene: &mut SceneWriter<'_>,
-    ias: f32,
-    lo_kt: f32,
-    hi_kt: f32,
-    x: f32,
-    w: f32,
-    color: Rgba8,
-) -> Result<(), SceneError> {
-    let y_top = (CENTER_Y - (hi_kt - ias) * PX_PER_KT).max(SPEED_TAPE_TOP);
-    let y_bot = (CENTER_Y - (lo_kt - ias) * PX_PER_KT).min(335.0);
-    if y_bot > y_top {
-        scene.fill_color(color)?;
-        scene.rect(PaintMode::Fill, x, y_top, w, y_bot - y_top)?;
-    }
-    Ok(())
-}
+/// The foot both tapes end at. The readout boxes hang from it, and a
+/// cue drawn past it would mark a value the tape beside it is not
+/// showing.
+const TAPE_BOTTOM: f32 = 335.0;
 
 /// Geometry of a pointed tape readout: the rectangular body spans
 /// `far_x`..`near_x`, the tip at `tip_x` points toward the tape, and
@@ -311,200 +128,4 @@ fn fitted_text_size(geo: &PointedBox, chars: usize) -> f32 {
         scale = scale.min((body_right - geo.text_x) / right_need);
     }
     geo.preferred_size * scale.max(0.0)
-}
-
-/// Right-edge altitude tape with selected-altitude bug and baro box.
-/// The tape carries its reference label (REL amber, BARO/STD/MSL/AGL
-/// white, RED for an unknown reference) so a local-relative height can
-/// never read as barometric altitude; the bug and selection readout
-/// render only when the selection's reference class matches.
-pub fn altitude_tape(scene: &mut SceneWriter<'_>, data: &PanelData) -> Result<(), SceneError> {
-    let alt = data.altitude.value_ft;
-    scene.fill_color(palette::TAPE_BG)?;
-    scene.rect(PaintMode::Fill, 390.0, 0.0, 90.0, 335.0)?;
-
-    if alt.status.shows_value() {
-        scene.save()?;
-        scene.clip_rect(390.0, 0.0, 90.0, 335.0)?;
-        scene.stroke(palette::WHITE, 2.0)?;
-        scene.fill_color(palette::WHITE)?;
-        let lo = ((alt.value - 155.0) / 20.0) as i32;
-        let hi = ((alt.value + 155.0) / 20.0) as i32;
-        for step in lo..=hi {
-            let ft = step * 20;
-            let y = CENTER_Y - (ft as f32 - alt.value) * PX_PER_FT;
-            scene.line(390.0, y, 400.0, y)?;
-            if step.rem_euclid(5) == 0 {
-                let label = fmt_label!(12, "{ft}");
-                scene.text_attributed(
-                    altitude_claim(data),
-                    408.0,
-                    y,
-                    18.0,
-                    Anchor::MIDDLE_LEFT,
-                    label.as_str(),
-                )?;
-            }
-        }
-        if let (true, Some(sel_m)) = (data.altitude.bug_compatible, data.selections.altitude_sel_m)
-        {
-            let sel_ft = sel_m * indicate_instrument_state::units::M_TO_FT;
-            let y = (CENTER_Y - (sel_ft - alt.value) * PX_PER_FT).clamp(4.0, 331.0);
-            scene.fill_color(palette::CYAN)?;
-            scene.polygon(
-                PaintMode::Fill,
-                &[
-                    [390.0, y - 8.0],
-                    [398.0, y - 8.0],
-                    [398.0, y - 3.0],
-                    [393.0, y],
-                    [398.0, y + 3.0],
-                    [398.0, y + 8.0],
-                    [390.0, y + 8.0],
-                ],
-            )?;
-        }
-        scene.restore()?;
-    } else {
-        scene.fill_color(palette::GREY)?;
-        scene.text(435.0, 130.0, 16.0, Anchor::CENTER, "ALT")?;
-    }
-
-    let text = fmt_label!(12, "{}", libm::roundf(alt.value / 10.0) as i64 * 10);
-    pointed_readout(
-        scene,
-        altitude_claim(data),
-        alt,
-        text.as_str(),
-        &ALT_READOUT,
-    )?;
-    reference_label(scene, data)?;
-    baro_and_sel_boxes(scene, data)?;
-    Ok(())
-}
-
-/// The group an altitude value derives from under the declared class:
-/// local-relative altitude is kinematic; every other class rides the
-/// air-data group's stamp (the altitude group only qualifies the
-/// datum), and its claim must say so or a barometric altitude would be
-/// refused as fabricated under kinematics withholding.
-fn altitude_claim(data: &PanelData) -> u8 {
-    match data.altitude.class {
-        AltitudeClass::LocalRelative => GroupId::Kinematics.to_u8(),
-        _ => GroupId::Air.to_u8(),
-    }
-}
-
-/// The altitude reference label under the value box. REL is amber —
-/// simulator-relative height demands attention — and an unknown wire
-/// reference is red beside its failed tape.
-fn reference_label(scene: &mut SceneWriter<'_>, data: &PanelData) -> Result<(), SceneError> {
-    let class = data.altitude.class;
-    scene.fill_color(match class {
-        AltitudeClass::LocalRelative => safety::CAUTION_AMBER,
-        AltitudeClass::Unknown => safety::FAILURE_RED,
-        _ => palette::WHITE,
-    })?;
-    scene.text(442.0, 222.0, 12.0, Anchor::CENTER, class.label())?;
-    Ok(())
-}
-
-/// Setting and selection boxes. The setting readout states whether the
-/// shown setting is applied: a barometric tape shows the applied value
-/// in cyan, a pressure tape shows STD, and every other reference shows
-/// the setting prefixed SET in grey — visibly not applied to the tape.
-/// A selected/applied disagreement adds the amber BARO SEL flag.
-fn baro_and_sel_boxes(scene: &mut SceneWriter<'_>, data: &PanelData) -> Result<(), SceneError> {
-    let baro = data.baro_hpa;
-    let (text, color) = match data.altitude.class {
-        AltitudeClass::BaroIndicated => (fmt_label!(12, "{:.0}", baro.value), palette::CYAN),
-        AltitudeClass::Pressure => (fmt_label!(12, "STD"), palette::CYAN),
-        _ => (fmt_label!(12, "SET {:.0}", baro.value), palette::GREY),
-    };
-    status_paint::readout_box(
-        scene,
-        GroupId::Air.to_u8(),
-        390.0,
-        335.0,
-        90.0,
-        25.0,
-        text.as_str(),
-        color,
-        16.0,
-        baro.status,
-    )?;
-    if data.altitude.setting_mismatch {
-        status_paint::draw_flag(scene, 435.0, 330.0, "BARO SEL")?;
-    }
-    match (data.altitude.bug_compatible, data.selections.altitude_sel_m) {
-        (true, Some(sel_m)) => {
-            let sel_ft = sel_m * indicate_instrument_state::units::M_TO_FT;
-            let text = fmt_label!(12, "{}", libm::roundf(sel_ft) as i64);
-            scene.fill_color(palette::BOX_BG)?;
-            scene.stroke(palette::GREY, 1.5)?;
-            scene.rect(PaintMode::FillStroke, 390.0, 0.0, 90.0, 24.0)?;
-            scene.fill_color(palette::CYAN)?;
-            scene.text_attributed(
-                GroupId::Selections.to_u8(),
-                435.0,
-                12.0,
-                18.0,
-                Anchor::CENTER,
-                text.as_str(),
-            )?;
-        }
-        (false, Some(_)) => {
-            // A selection in an incompatible reference never renders as
-            // a plausible number; the amber marker says why it is gone.
-            scene.fill_color(safety::CAUTION_AMBER)?;
-            scene.text(435.0, 12.0, 14.0, Anchor::CENTER, "SEL REF")?;
-        }
-        (_, None) => {}
-    }
-    Ok(())
-}
-
-/// Vertical-speed bar at the right edge of the altitude tape.
-pub fn vsi(scene: &mut SceneWriter<'_>, data: &PanelData) -> Result<(), SceneError> {
-    let v = data.vsi_fpm;
-    scene.stroke(palette::GREY, 1.0)?;
-    for dy in [-120.0f32, -60.0, 60.0, 120.0] {
-        scene.line(466.0, CENTER_Y + dy, 474.0, CENTER_Y + dy)?;
-    }
-    if !v.status.shows_value() {
-        // Dashes rather than an empty strip. Every other readout on this
-        // panel says "I have no value" out loud, and a blank scale reads
-        // as zero vertical speed to anyone scanning it — which is the
-        // fabrication this signal's own validity exists to prevent. The
-        // dash path stays unclaimed, like the others: it is the honest
-        // degraded display, not a value derived from a withheld group.
-        scene.fill_color(palette::WHITE)?;
-        scene.text(452.0, CENTER_Y, 12.0, Anchor::CENTER, "---")?;
-        return Ok(());
-    }
-    // ±1500 fpm full scale over 180 px.
-    let len = (v.value / 1500.0 * 180.0).clamp(-170.0, 170.0);
-    scene.fill_color(palette::MAGENTA)?;
-    if len >= 0.0 {
-        scene.rect(PaintMode::Fill, 466.0, CENTER_Y - len, 8.0, len.max(1.0))?;
-    } else {
-        scene.rect(PaintMode::Fill, 466.0, CENTER_Y, 8.0, -len)?;
-    }
-    if v.value.abs() >= 100.0 {
-        // Clamped clear of the selected-altitude box above and the baro
-        // box below: at full scale the label parks at the strip edge
-        // instead of sliding its ink under a readout box.
-        let tip_y = (CENTER_Y - len).clamp(32.0, 328.0);
-        let label = fmt_label!(12, "{}", libm::roundf(v.value / 50.0) as i64 * 50);
-        scene.fill_color(palette::WHITE)?;
-        scene.text_attributed(
-            GroupId::Kinematics.to_u8(),
-            452.0,
-            tip_y,
-            12.0,
-            Anchor::CENTER,
-            label.as_str(),
-        )?;
-    }
-    Ok(())
 }
